Split ExecutionCommitteeMinimumApproval component
